+++ src/element.rs
@@ -1,6 +1,11 @@
 use std::any::TypeId;
+use std::marker::PhantomData;
+use std::mem::ManuallyDrop;
+use std::ptr::NonNull;
 use crate::any_value::{AnyValue, AnyValueCloneable, AnyValueMut, clone_into, Unknown};
 use crate::{AnyVec, refs};
+use crate::any_vec_raw::AnyVecRaw;
+use crate::any_vec_ptr::{AnyVecPtr, IAnyVecPtr, IAnyVecRawPtr};
 use crate::traits::{Cloneable, Trait};
 
 /// Pointer to [`AnyVec`] element.
@@ -13,12 +18,23 @@ use crate::traits::{Cloneable, Trait};
 /// `Element` have it's own implementation of `downcast_` family (which return `&'a T`, instead of `&T`).
 /// This is done, so you don't have to keep ElementRef/Mut alive, while casting to concrete type.
 /// [`AnyValueMut`] implemented too - for the sake of interface compatibility.
-pub struct Element<'a, Traits: ?Sized + Trait>{
-    pub(crate) any_vec: &'a AnyVec<Traits>,
-    pub(crate) element: *const u8
+pub struct Element<'a, AnyVecPtr: IAnyVecRawPtr>{
+    any_vec_ptr: AnyVecPtr,
+    element: NonNull<u8>,
+    phantom: PhantomData<&'a mut AnyVecRaw>
 }
 
-impl<'a, Traits: ?Sized + Trait> Element<'a, Traits>{
+impl<'a, AnyVecPtr: IAnyVecRawPtr> Element<'a, AnyVecPtr>{
+    #[inline]
+    pub fn new(any_vec_ptr: AnyVecPtr, element: NonNull<u8>) -> Self {
+        Self{any_vec_ptr, element, phantom: PhantomData}
+    }
+
+    #[inline]
+    fn any_vec_raw(&self) -> &'a AnyVecRaw{
+        unsafe { self.any_vec_ptr.any_vec_raw().as_ref() }
+    }
+
     #[inline]
     pub fn downcast_ref<T: 'static>(&self) -> Option<&'a T>{
         if self.value_typeid() != TypeId::of::<T>(){
@@ -48,40 +64,59 @@ impl<'a, Traits: ?Sized + Trait> Element<'a, Traits>{
     }
 }
 
-impl<'a, Traits: ?Sized + Trait> AnyValue for Element<'a, Traits>{
+// TODO: specialized drop for known type
+impl<'a, AnyVecPtr: IAnyVecRawPtr> Drop for Element<'a, AnyVecPtr>{
+    #[inline]
+    fn drop(&mut self) {
+        if let Some(drop_fn) = self.any_vec_raw().drop_fn(){
+            (drop_fn)(self.element.as_ptr(), 1);
+        }
+    }
+}
+
+impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyValue for Element<'a, AnyVecPtr>{
     type Type = Unknown;
 
+    #[inline]
     fn value_typeid(&self) -> TypeId {
-        self.any_vec.element_typeid()
+        self.any_vec_raw().element_typeid()
     }
 
+    #[inline]
     fn size(&self) -> usize {
-        self.any_vec.element_layout().size()
+        self.any_vec_raw().element_layout().size()
     }
 
+    #[inline]
     fn bytes(&self) -> *const u8 {
-        self.element
+        self.element.as_ptr()
     }
 }
 
-impl<'a, Traits: ?Sized + Trait> AnyValueMut for Element<'a, Traits>{}
+impl<'a, AnyVecPtr: IAnyVecRawPtr> AnyValueMut for Element<'a, AnyVecPtr>{}
 
-impl<'a, Traits: ?Sized + Cloneable + Trait> AnyValueCloneable for Element<'a, Traits>{
+impl<'a, Traits: ?Sized + Cloneable + Trait>
+    AnyValueCloneable for Element<'a, AnyVecPtr<Traits>>
+{
     #[inline]
     unsafe fn clone_into(&self, out: *mut u8) {
-        clone_into(self, out, self.any_vec.clone_fn());
+        clone_into(self, out, self.any_vec_ptr.any_vec().as_ref().clone_fn());
     }
 }
 
-unsafe impl<'a, Traits: ?Sized + Send + Trait> Send for Element<'a, Traits>{}
-unsafe impl<'a, Traits: ?Sized + Sync + Trait> Sync for Element<'a, Traits>{}
+unsafe impl<'a, Traits: ?Sized + Send + Trait> Send for Element<'a, AnyVecPtr<Traits>>{}
+unsafe impl<'a, Traits: ?Sized + Sync + Trait> Sync for Element<'a, AnyVecPtr<Traits>>{}
 
 /// Reference to [`AnyVec`] element.
 ///
 /// Created by  [`AnyVec::get`].
-pub type ElementRef<'a, Traits> = refs::Ref<Element<'a, Traits>>;
+pub type ElementRef<'a, Traits> = refs::Ref<ManuallyDrop<
+    Element<'a, AnyVecPtr<Traits>>
+>>;
 
 /// Mutable reference to [`AnyVec`] element.
 ///
 /// Created by  [`AnyVec::get_mut`].
-pub type ElementMut<'a, Traits> = refs::Mut<Element<'a, Traits>>;
+pub type ElementMut<'a, Traits> = refs::Mut<ManuallyDrop<
+    Element<'a, AnyVecPtr<Traits>>
+>>;