+++ src/lib.rs
@@ -74,8 +74,10 @@
 
 mod any_vec;
 mod clone_type;
+mod any_vec_ptr;
 mod any_vec_raw;
 mod any_vec_typed;
+mod iter;
 
 pub use crate::any_vec::{AnyVec, AnyVecMut, AnyVecRef, SatisfyTraits, traits};
 pub use any_vec_typed::AnyVecTyped;