+++ src/ops/swap_remove.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 use std::ptr;
 use crate::copy_bytes_nonoverlapping;
 use crate::any_value::Unknown;
-use crate::ops::any_vec_ptr::IAnyVecRawPtr;
+use crate::any_vec_ptr::IAnyVecRawPtr;
 use crate::ops::temp::Operation;
 
 pub struct SwapRemove<'a, AnyVecPtr: IAnyVecRawPtr, T: 'static = Unknown>{